Refactor names, e.g. shell and home modules
